added puthex to print hexadecimal

+++ Project/ft_printf.h
@@ -6,16 +6,18 @@
 /*   By: abdnasse <abdnasse@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/27 18:44:42 by abdnasse          #+#    #+#             */
-/*   Updated: 2024/11/28 16:06:57 by abdnasse         ###   ########.fr       */
+/*   Updated: 2024/11/29 17:05:44 by abdnasse         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF
 # define FT_PRINTF
 
+#include <unistd.h>
+
 int ft_printf(const char *format, ...);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_puthex(char c, long n);
+int	ft_puthex(char c, unsigned int n);
 
 #endif
